requirements.txt
Django==3.0.5
web3==5.7.0
django-web3-auth==0.1.0
dj-database-url==0.5.0
django-cors-headers==3.2.1
django-extensions==2.2.9
whitenoise==5.0.1
psycopg2==2.8.5
rlp==1.2.0 
